@@ -3,7 +3,6 @@ import { environment} from "../../environments/environment";
 import {HttpClient, HttpErrorResponse, HttpEvent, HttpResponse} from "@angular/common/http";
 import {Observable} from "rxjs";
 import {User} from "../models/user";
-import {JwtHelperService, JwtInterceptor} from "@auth0/angular-jwt";
 import {CustomHttpResponse} from "../Http-Response/Custom-http-response";
 
 @Injectable({
@@ -15,28 +14,28 @@ export class UserService {
   constructor(private http: HttpClient) { }
 
   public getUsers(): Observable<User[] | HttpErrorResponse>{
-    return  this.http.get<User[]>(`${this.host}/user/list`);
+    return  this.http.get<User[]>(`${this.host}/api/archive/user/list`);
   }
   public addUser(formData: FormData): Observable<any>{
-    return  this.http.post<User>(`${this.host}/user/add`,formData);
+    return  this.http.post<User>(`${this.host}/api/archive/user/add`,formData);
   }
-  public updateUser(formData: FormData): Observable<User | HttpErrorResponse>{
-    return  this.http.put<User>(`${this.host}/user/update`,formData);
+  public updateUser(formData: FormData): Observable<any>{
+    return  this.http.put<User>(`${this.host}/api/archive/user/update`,formData);
   }
   public resetPassword(email: string): Observable<CustomHttpResponse | HttpErrorResponse>{
-    return  this.http.get<CustomHttpResponse>(`${this.host}/user/resetPassword/${email}`);
+    return  this.http.get<CustomHttpResponse>(`${this.host}/api/archive/user/resetPassword/${email}`);
   }
 
   public updateProfilrImage(formData: FormData): Observable<HttpEvent<User> | HttpErrorResponse>{
-    return  this.http.post<User>(`${this.host}/user/updateProfileImage`,formData,
+    return  this.http.post<User>(`${this.host}/api/archive/user/updateProfileImage`,formData,
       {reportProgress: true,
       observe: 'events'
       });
   }
   public deleteUser(userId: number): Observable<CustomHttpResponse | HttpErrorResponse>{
-    return  this.http.delete<CustomHttpResponse>(`${this.host}/user/delete/${userId}`);
+    return  this.http.delete<CustomHttpResponse>(`${this.host}/api/archive/user/delete/${userId}`);
   }
-  public addUsersToLocalCache(users: User[]): void{
+  public addUsersToLocalCache(users: User[] | HttpErrorResponse): void{
     localStorage.setItem('users',JSON.stringify(users));
   }
   public getUsersFromLocalCache(): User[] | null {
@@ -57,12 +56,12 @@ export class UserService {
     formData.append('role',user.role);
     formData.append('profileImage',profileImage);
     formData.append('isActive',JSON.stringify(user.active));
-    formData.append('isNomLocked',JSON.stringify(user.notLocked));
+    formData.append('isNotLocked',JSON.stringify(user.notLocked));
     return formData;
   }
 
  /* public deleteUsers(userId: number): Observable<any> {
-    return this.http.delete<User>(`${this.host}/user/delete/${userId}`, { observe: 'response' })
+    return this.http.delete<User>(`${this.host}/api/archive/user/delete/${userId}`, { observe: 'response' })
       .pipe(
         catchError(this.handleError)
       );
