@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot} from '@angular/router';
+import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree} from '@angular/router';
 import {AuthenticationService} from "../GlobaleServices/authentication.service";
 import {NotificationService} from "../GlobaleServices/notification.service";
 import {NotificationType} from "../Enumerations/notification-type.enum";
@@ -14,16 +14,16 @@ export class AuthenticationGuard implements CanActivate {
   }
   canActivate(
     route: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot):  boolean  {
+    state: RouterStateSnapshot):  boolean | UrlTree {
     return this.isUserLoggedIn();
   }
 
   private isUserLoggedIn(): boolean{
     if(this.authenticationService.isUserLoggedIn()){
-
+      return true;
     }
     this.router.navigate(['/login'])
-    this.notificationService.notifyShow(NotificationType.ERREUR,
+    this.notificationService.notify(NotificationType.ERREUR,
       'You need to log in to access this page'.toUpperCase());
     return false;
   }
